Implementing StarRating API
Also, Implementing Response Entity to addRecipe method for easier debugging

+++ foodApp/src/main/java/com/example/foodApp/Backend/Fuzzy.java
@@ -109,9 +109,9 @@ public class Fuzzy {
             adjustIngredientQuantities(recipe, servingSize);
 
             // Generate score for the recipe
-            float starRating = recipeManager.getStarRating(recipe.getId());
+            float starRating = (float) recipeManager.getStarRating(recipe.getId());
             float difficultyRating = recipeManager.getDifficultyRating(recipe.getId());
-            double score = generateRecipeScore(recipe, starRating, difficultyRating, pantry);
+            double score = generateRecipeScore(recipe,  starRating, difficultyRating, pantry);
 
             // Create a map to store score and recipe
             Map<Double, Recipe> scoreRecipeMap = new HashMap<>();
